% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ayu.map}
\alias{ayu.map}
\title{Ayu colour scheme}
\format{A data frame containing all the colours used in the palette:
\itemize{
  \item V1: Red value
  \item V2: Green value
  \item V3: Blue value
  \item option: It is intended to be a general option for choosing the specific colour palette.
}}
\usage{
ayu.map
}
\description{
A dataset containing some colour palettes inspired on the Ayu theme
}
\keyword{datasets}
