% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ayu_palettes}
\alias{ayu_palettes}
\title{Available Palettes.}
\format{A list containing all palettes color codes.}
\usage{
ayu_palettes
}
\description{
This list contains all the available palettes in the 'ayu' package.
}
\keyword{datasets}
