% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_palette.R
\name{display_palette}
\alias{display_palette}
\title{Display Palette}
\usage{
display_palette(pal, xlab = "", ylab = "")
}
\arguments{
\item{pal}{Array of colours in hex code}

\item{xlab}{Label in the x axis}

\item{ylab}{Label in the y axis}
}
\description{
Display Palette
}
