% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_continuous_palette.R
\name{generate_continuous_palette}
\alias{generate_continuous_palette}
\title{Generate Continuous Palette}
\usage{
generate_continuous_palette(colour, contrast = 0.3, n = 7)
}
\arguments{
\item{colour}{hex code}

\item{contrast}{number from 0 to 1}

\item{n}{number of colours}
}
\description{
Generate Continuous Palette
}
